feat: [#173103563] Move all bonus routes under bonusVacanze directory and remove bottom bar from all bonus routes (#1871)
* [#173103563] Move all bonus routes under bonusVacanze directory and remove bottom bar from all bonus routes

* Removed log

* Refactoring and restores the stack navigator for bonus
