+++ ts/features/bonusVacanze/navigation/action.ts
@@ -1,4 +1,7 @@
 import { NavigationActions } from "react-navigation";
+import { InferNavigationParams } from "../../../types/react";
+import ActiveBonusScreen from "../screens/ActiveBonusScreen";
+import BonusInformationScreen from "../screens/BonusInformationScreen";
 import BONUSVACANZE_ROUTES from "./routes";
 
 export const navigateToBonusEligibilityLoading = () =>
@@ -24,6 +27,32 @@ export const navigateToActivateBonus = () =>
 export const navigateToTimeoutEligibilityCheck = () =>
   NavigationActions.navigate({
     routeName: BONUSVACANZE_ROUTES.ELIGIBILITY.TIMEOUT
+  });
+
+export const navigateToAvailableBonusScreen = () =>
+  NavigationActions.navigate({
+    routeName: BONUSVACANZE_ROUTES.BONUS_AVAILABLE_LIST
+  });
+
+export const navigateToBonusRequestInformation = (
+  params?: InferNavigationParams<typeof BonusInformationScreen>
+) =>
+  NavigationActions.navigate({
+    routeName: BONUSVACANZE_ROUTES.BONUS_REQUEST_INFORMATION,
+    params
+  });
+
+export const navigateToBonusTosScreen = () =>
+  NavigationActions.navigate({
+    routeName: BONUSVACANZE_ROUTES.BONUS_TOS_SCREEN
+  });
+
+export const navigateToBonusActiveDetailScreen = (
+  params?: InferNavigationParams<typeof ActiveBonusScreen>
+) =>
+  NavigationActions.navigate({
+    routeName: BONUSVACANZE_ROUTES.BONUS_ACTIVE_DETAIL_SCREEN,
+    params
   });
 
 export const navigateToActivationLoading = () =>
