+++ ts/features/bonusVacanze/navigation/routes.ts
@@ -1,6 +1,11 @@
 const BONUSVACANZE_ROUTES = {
   MAIN: "MAIN",
 
+  BONUS_AVAILABLE_LIST: "BONUS_AVAILABLE_LIST",
+  BONUS_REQUEST_INFORMATION: "BONUS_REQUEST_INFORMATION",
+  BONUS_TOS_SCREEN: "BONUS_TOS_SCREEN",
+  BONUS_ACTIVE_DETAIL_SCREEN: "BONUS_ACTIVE_DETAIL_SCREEN",
+
   // Grouping the "Bonus Eligibility Verification" sub-phase
   ELIGIBILITY: {
     CHECK_LOADING: "BONUS_ELIGIBILITY_CHECK_LOADING",
