+++ ts/navigation/WalletNavigator.ts
@@ -2,10 +2,6 @@ import { createStackNavigator } from "react-navigation";
 import { bonusVacanzeEnabled } from "../config";
 import BonusVacanzeNavigator from "../features/bonusVacanze/navigation/navigator";
 import BONUSVACANZE_ROUTES from "../features/bonusVacanze/navigation/routes";
-import AcceptTosBonusScreen from "../features/bonusVacanze/screens/AcceptTosBonusScreen";
-import ActiveBonusScreen from "../features/bonusVacanze/screens/ActiveBonusScreen";
-import AvailableBonusScreen from "../features/bonusVacanze/screens/AvailableBonusScreen";
-import BonusInformationScreen from "../features/bonusVacanze/screens/BonusInformationScreen";
 import AddCardScreen from "../screens/wallet/AddCardScreen";
 import AddPaymentMethodScreen from "../screens/wallet/AddPaymentMethodScreen";
 import ConfirmCardDetailsScreen from "../screens/wallet/ConfirmCardDetailsScreen";
@@ -82,18 +78,6 @@ const baseRouteConfigMap = {
 const bonusVacanzeConfigMap = {
   [BONUSVACANZE_ROUTES.MAIN]: {
     screen: BonusVacanzeNavigator
-  },
-  [ROUTES.BONUS_REQUEST_INFORMATION]: {
-    screen: BonusInformationScreen
-  },
-  [ROUTES.BONUS_AVAILABLE_LIST]: {
-    screen: AvailableBonusScreen
-  },
-  [ROUTES.BONUS_TOS_SCREEN]: {
-    screen: AcceptTosBonusScreen
-  },
-  [ROUTES.BONUS_ACTIVE_DETAIL_SCREEN]: {
-    screen: ActiveBonusScreen
   }
 };
