+++ ts/navigation/routes.ts
@@ -43,12 +43,6 @@ const ROUTES = {
   WALLET_CONFIRM_CARD_DETAILS: "WALLET_CONFIRM_CARD_DETAILS",
   WALLET_CHECKOUT_3DS_SCREEN: "WALLET_CHECKOUT_3DS_SCREEN",
 
-  // Bonus
-  BONUS_AVAILABLE_LIST: "BONUS_AVAILABLE_LIST",
-  BONUS_REQUEST_INFORMATION: "BONUS_REQUEST_INFORMATION",
-  BONUS_TOS_SCREEN: "BONUS_TOS_SCREEN",
-  BONUS_ACTIVE_DETAIL_SCREEN: "BONUS_ACTIVE_DETAIL_SCREEN",
-
   // Payment
   PAYMENT_SCAN_QR_CODE: "PAYMENT_SCAN_QR_CODE",
   PAYMENT_MANUAL_DATA_INSERTION: "PAYMENT_MANUAL_DATA_INSERTION",
